fix: only deal with a dynamic link once
If the caller method of respondWithDynamicLink is getDynamicLink, we want to use the timestamp logic to verify if that dynamic link was already dealt with, and we want to update the timestamp if it is a new dynamic link. If the method is called from the onDynamicLink event, we do not want to update the timestamp.